requirements.txt
flask
apispec
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
